requirements.txt
discord.py
requests
gtts
SpeechRecognition
python-dotenv
